app/requirements.txt
ipdb
# console stuff
click
requests

# django stuff
pyyaml
matplotlib

